# openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GHRI Foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers/auth/register:
    post:
      operationId: volRegister
      tags: [volunteer-auth]
      summary: Register as volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolAuthResponse"
  /volunteers/auth/login:
    post:
      operationId: volLogin
      tags: [volunteer-auth]
      summary: Login to volunteer portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolAuthResponse"
  /volunteers/auth/logout:
    post:
      operationId: volLogout
      tags: [volunteer-auth]
      summary: Logout from volunteer portal
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
  /volunteers/auth/me:
    get:
      operationId: volMe
      tags: [volunteer-auth]
      summary: Get current volunteer profile
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolUserProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VolUserProfile:
      type: object
      required: [id, email, name, role, status]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [volunteer, coordinator]
        status:
          type: string
          enum: [pending, active, inactive]
        avatarInitials:
          type: ["string", "null"]
        skills:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
    VolAuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/VolUserProfile"
...
    VolRegisterInput:
      type: object
      required: [email, password, name, consentedToTerms]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        skills:
          type: string
        availability:
          type: string
        consentedToTerms:
          type: boolean
    VolLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...